Refactored CPL names
