+++ src/lib/index.js
@@ -4,4 +4,4 @@ export * from './errors';
 export * from './isRTLLanguage';
 export * from './security';
 export * from './session';
-export * from './perceivedLatencyInstrumentation';
+export * from './perceived-latency-instrumentation';
