+++ src/lib/perceived-latency-instrumentation.js
